BROKEN - but I don't have more time right now
This is becoming rapidly unmaintainable. I'm going to have to slow down
my pace of development and write smaller tests. I've been working the
way I have because I have a concept and I want to get it off the ground.
I know though that this can't continue, and it's starting to look like
now is the time when I will have to begin slowing down.